move filedropbox to image-converter dir
